student_attendance_bot/requirements.txt
python-telegram-bot==21.0.1
apscheduler==3.10.4
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.1
